Added physical simulation avoidance feature
Whenever we apply an impulse to a limb object, based on the magnitude of the impulse, we ignore physical collisions for a brief period at the beginning of the physics simulation. This ensures that the limb can maintain the dynamics of the explosive moment while still achieving relatively accurate physical effects.

The duration of the physics avoidance is inversely proportional to the force (velocity). The system is configured with an overall force (velocity) range and a maximum avoidance time for the project. When applying an impact to a limb, it automatically calculates the required physics avoidance time.

Additionally, the system guarantees that any limb object will have at least one frame (at 60fps) of physics simulation avoidance time.